Adds an allocator class that is currently used in the Parquet reader to allow management of allocations
