+++ extension/parquet/parquetcli.cpp
@@ -38,8 +38,9 @@ int main(int argc, const char **argv) {
 
 	// the db instance and client context are not really required so we may remove them
 
+	Allocator allocator;
 	FileSystem fs;
-	ParquetReader reader(fs.OpenFile(filename, FileFlags::FILE_FLAGS_READ));
+	ParquetReader reader(allocator, fs.OpenFile(filename, FileFlags::FILE_FLAGS_READ));
 
 	// only return columns first_name and last_name
 	std::vector<column_t> column_ids;
